add loading indicator and split shortcuts
