+++ packages/insomnia/src/index.html
@@ -9,12 +9,19 @@
 
 <body>
   <div id="root">
-    <div style="
+    <div
+      id="app-loading-indicator"
+      style="
+        position: fixed;
+        top: 0;
+        left: 0;
         display: flex;
         align-items: center;
         justify-content: center;
-        height:100%;
-      ">
+        height: 100%;
+        width: 100%;
+      "
+    >
       <img src="./ui/images/insomnia-logo.svg" alt="Insomnia">
     </div>
   </div>
